pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sat"
version = "0.1.0"
description = "Structured Analytic Techniques - CIA Tradecraft Primer-based analysis tool"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15.0",
    "rich>=13.0",
    "pydantic>=2.11",
    "anthropic>=0.45",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "markdown>=3.5",
    "mcp[cli]>=1.0",
]

[project.optional-dependencies]
openai = ["openai>=1.60"]
gemini = ["google-genai>=1.0"]
docling = ["docling>=2.0"]
api = ["fastapi>=0.115", "uvicorn[standard]>=0.34", "websockets>=14.0"]
all = ["openai>=1.60", "google-genai>=1.0", "docling>=2.0", "fastapi>=0.115", "uvicorn[standard]>=0.34", "websockets>=14.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
]

[project.scripts]
sat = "sat.cli:app"
sat-mcp = "sat.mcp_server:main"
sat-api = "sat.api.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sat"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src", "."]
markers = [
    "integration: tests that call real LLM APIs (deselected by default, run with -m integration)",
]
addopts = "-m 'not integration'"

[tool.ruff]
line-length = 100
target-version = "py311"
